fix login and intent parcelable detail
