+++ app/src/main/java/com/androidwave8/dailygadget/App.kt
@@ -1,24 +1,27 @@
 package com.androidwave8.dailygadget
 
+import android.annotation.SuppressLint
 import android.app.Application
 import android.content.Context
 import androidx.room.Room
-import com.androidwave8.dailygadget.data.db.Database
+import com.androidwave8.dailygadget.data.db.DB
 
-class App: Application() {
+class App : Application() {
 
     companion object {
+        @SuppressLint("StaticFieldLeak")
         lateinit var context: Context
-        lateinit var dbUser: Database
+        lateinit var db: DB
+        //lateinit var dbUser: Database
     }
 
     override fun onCreate() {
         super.onCreate()
         context = applicationContext
 
-        dbUser = Room.databaseBuilder(
+        db = Room.databaseBuilder(
             context.applicationContext,
-            Database::class.java, "user"
+            DB::class.java, "DailyGadget"
         ).build()
     }
 }
