+++ app/src/main/java/com/androidwave8/dailygadget/data/ui/home/HomeAdapter.kt
@@ -5,7 +5,7 @@ import android.os.Parcelable
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.androidwave8.dailygadget.utils.App
+import com.androidwave8.dailygadget.App
 import com.androidwave8.dailygadget.data.db.Gadget
 import com.androidwave8.dailygadget.databinding.ListGadgetItemBinding
 import com.bumptech.glide.Glide
