+++ app/src/main/java/com/androidwave8/dailygadget/utils/SharePref.kt
@@ -1,9 +1,11 @@
 package com.androidwave8.dailygadget.utils
 
 import android.content.Context
+import com.androidwave8.dailygadget.App
 
 object SharePref {
 
+//    private val pref = App.context.getSharedPreferences("myPref", Context.MODE_PRIVATE)
     private val pref = App.context.getSharedPreferences("myPref", Context.MODE_PRIVATE)
 
     var isLogin: Boolean
